Changed the Poisson noise in the spectra module
Changed the lambda parameter in the Poisson noise generator. A value of 1000 was used, but this might be changed to more accurately reflect the noise characteristics of our instrument (see lines ~661 in spectra.py). The noise value is then divided by the value of lambda and multiplied by the signal_noise ratio.
The random signal_noise ratio in the creator needed to then be re-scaled (see lines ~ 165 in creator.py)