Changed noise calculation
The previously implemented calculation of the poisson noise was redefined such that the output S/N ratio resembles the actual S/N ratio in the spectrum (It is not exact due to normalization of the spectrum).

In accordance with these changes, the methods for writing the information about the simulation parameters during plotting were changed.